pyproject.toml
[project]
name = "regex-assistant"
version = "0.1.0"
description = "Simple regex assistant based on LLM"

readme = "README.md"

requires-python = ">=3.14.0,<4"

dependencies = [
    "fastapi>=0.115.8, <0.116",
    "uvicorn>=0.34.0,<0.35",
    "openai>=2.6.1, <3",
    "python-dotenv>=1.2.1, <2",
]